Update MedCY.036.001.xml
Deniz, from now on, majuscule will be included in the first sentence. Do not make any changes for it — only in cases where you reopen the file for further corrections. The word Θεός, Κύριος will be indexed as a lemma. We have revised certain aspects of the letter descriptions. Look at the changes I will make to a few of them and follow them accordingly.