Update MedCY.036.002.xml
Also, for the Arakas records I forgot to tell you to add also G. Apeitos on the text constituted from. In cases where you open for any reason records from Arakas, please remember to add it. Here according to your mark up in diplomatic the accents and breathings are systematic